+++ ICSharpCode.CodeConverter/CSharp/NodesVisitor.cs
@@ -34,6 +34,7 @@ namespace ICSharpCode.CodeConverter.CSharp
             private uint failedMemberConversionMarkerCount;
             private HashSet<string> _extraUsingDirectives = new HashSet<string>();
             private static readonly Type ExtensionAttributeType = typeof(ExtensionAttribute);
+            private readonly TypeConversionAnalyzer _typeConversionAnalyzer;
             public CommentConvertingNodesVisitor TriviaConvertingVisitor { get; }
 
             private CommonConversions CommonConversions { get; }
@@ -45,6 +46,7 @@ namespace ICSharpCode.CodeConverter.CSharp
                 TriviaConvertingVisitor = new CommentConvertingNodesVisitor(this);
                 _createConvertMethodsLookupByReturnType = CreateConvertMethodsLookupByReturnType(semanticModel);
                 CommonConversions = new CommonConversions(semanticModel, TriviaConvertingVisitor);
+                _typeConversionAnalyzer = new TypeConversionAnalyzer(semanticModel, csCompilation, CommonConversions, _extraUsingDirectives);
                 _queryConverter = new QueryConverter(CommonConversions, TriviaConvertingVisitor);
                 _additionalInitializers = new AdditionalInitializers();
             }
@@ -1395,8 +1397,7 @@ namespace ICSharpCode.CodeConverter.CSharp
                 }
                 return SyntaxFactory.Argument(
                     node.IsNamed ? SyntaxFactory.NameColon((IdentifierNameSyntax)node.NameColonEquals.Name.Accept(TriviaConvertingVisitor)) : null,
-                    token,
-                    AddExplicitConversion(node.Expression, (ExpressionSyntax)node.Expression.Accept(TriviaConvertingVisitor))
+                    token, _typeConversionAnalyzer.AddExplicitConversion(node.Expression, (ExpressionSyntax)node.Expression.Accept(TriviaConvertingVisitor))
                 );
             }
 
@@ -1597,8 +1598,8 @@ namespace ICSharpCode.CodeConverter.CSharp
                     }
                 }
 
-                var lhs = AddExplicitConversion(node.Left, (ExpressionSyntax)node.Left.Accept(TriviaConvertingVisitor));
-                var rhs = AddExplicitConversion(node.Right, (ExpressionSyntax)node.Right.Accept(TriviaConvertingVisitor));
+                var lhs = _typeConversionAnalyzer.AddExplicitConversion(node.Left, (ExpressionSyntax)node.Left.Accept(TriviaConvertingVisitor));
+                var rhs = _typeConversionAnalyzer.AddExplicitConversion(node.Right, (ExpressionSyntax)node.Right.Accept(TriviaConvertingVisitor));
 
                 if (node.IsKind(VBasic.SyntaxKind.ConcatenateExpression)) {
                     var stringType = _semanticModel.Compilation.GetTypeByMetadataName("System.String");
@@ -1608,7 +1609,7 @@ namespace ICSharpCode.CodeConverter.CSharp
                         lhsTypeInfo.ConvertedType.SpecialType != SpecialType.System_String &&
                         rhsTypeInfo.Type.SpecialType != SpecialType.System_String &&
                         rhsTypeInfo.ConvertedType.SpecialType != SpecialType.System_String) {
-                        lhs = AddExplicitConvertTo(node.Left, lhs, stringType);
+                        lhs = _typeConversionAnalyzer.AddExplicitConvertTo(node.Left, lhs, stringType);
                     }
                 }
 
@@ -1623,89 +1624,12 @@ namespace ICSharpCode.CodeConverter.CSharp
                 var op = SyntaxFactory.Token(CSharpUtil.GetExpressionOperatorTokenKind(kind));
 
                 var csBinExp = SyntaxFactory.BinaryExpression(kind, lhs, op, rhs);
-                var convertedNode = AddExplicitConversion(node, csBinExp, addParenthesisIfNeeded: true);
+                var convertedNode = _typeConversionAnalyzer.AddExplicitConversion(node, csBinExp, addParenthesisIfNeeded: true);
                 if (convertedNode == csBinExp) {
                     return CommonConversions.ParenthesizeIfPrecedenceCouldChange(node, csBinExp);
                 } else {
                     return convertedNode;
                 }
-            }
-
-            private ExpressionSyntax AddExplicitConversion(VBSyntax.ExpressionSyntax vbNode, ExpressionSyntax csNode, bool addParenthesisIfNeeded = false)
-            {
-                var typeInfo = _semanticModel.GetTypeInfo(vbNode);
-                var vbType = typeInfo.Type;
-                var vbConvertedType = typeInfo.ConvertedType;
-                if (vbType is null || vbConvertedType is null) {
-                    return csNode;
-                }
-
-                var vbCompilation = _semanticModel.Compilation as VBasic.VisualBasicCompilation;
-                var vbConversion = vbCompilation.ClassifyConversion(vbType, vbConvertedType);
-
-                var csType = _csCompilation.GetTypeByMetadataName(vbType.GetFullMetadataName());
-                var csConvertedType = _csCompilation.GetTypeByMetadataName(vbConvertedType.GetFullMetadataName());
-
-                if (csType is null || csConvertedType is null) {
-                    return csNode;
-                }
-
-                var csConversion = _csCompilation.ClassifyConversion(csType, csConvertedType);
-
-                bool insertConvertTo = false;
-                bool insertCast = false;
-
-                bool isConvertToString = vbConversion.IsString && vbConvertedType.SpecialType == SpecialType.System_String;
-                bool isArithmetic = vbNode.IsKind(VBasic.SyntaxKind.AddExpression, VBasic.SyntaxKind.SubtractExpression,
-                                                  VBasic.SyntaxKind.MultiplyExpression, VBasic.SyntaxKind.DivideExpression,
-                                                  VBasic.SyntaxKind.IntegerDivideExpression);
-                if (!csConversion.Exists) {
-                    insertConvertTo = isConvertToString || vbConversion.IsNarrowing;
-                } else if (vbConversion.IsWidening && vbConversion.IsNumeric && csConversion.IsImplicit && csConversion.IsNumeric) {
-                    // Safe overapproximation: A cast is really only needed to help resolve the overload for the operator/method used.
-                    // e.g. When VB "&" changes to C# "+", there are lots more overloads available that implicit casts could match.
-                    // e.g. sbyte * ulong uses the decimal * operator in VB. In C# it's ambiguous - see ExpressionTests.vb "TestMul".
-                    insertCast = true;
-                } else if (csConversion.IsExplicit && vbConversion.IsNumeric && vbType.TypeKind != TypeKind.Enum) {
-                    insertConvertTo = true;
-                } else if (isArithmetic) {
-                    var arithmeticConversion = vbCompilation.ClassifyConversion(vbConvertedType, vbCompilation.GetTypeByMetadataName("System.Int32"));
-                    if (arithmeticConversion.IsWidening && !arithmeticConversion.IsIdentity) {
-                        insertConvertTo = true;
-                    }
-                }
-
-                if (insertConvertTo) {
-                    return AddExplicitConvertTo(vbNode, csNode, vbConvertedType);
-                } else if (insertCast) {
-                    var typeName = CommonConversions.ToCsTypeSyntax(vbConvertedType, vbNode);
-                    if (csNode is CastExpressionSyntax cast && cast.Type.IsEquivalentTo(typeName)) {
-                        return csNode;
-                    }
-
-                    csNode = addParenthesisIfNeeded ? (ExpressionSyntax)CommonConversions.ParenthesizeIfPrecedenceCouldChange(vbNode, csNode) : csNode;
-                    return SyntaxFactory.CastExpression(typeName, csNode);
-                }
-
-                return csNode;
-            }
-
-            private ExpressionSyntax AddExplicitConvertTo(VBSyntax.ExpressionSyntax vbNode, ExpressionSyntax csNode, ITypeSymbol type)
-            {
-                var displayType = type.ToMinimalDisplayString(_semanticModel, vbNode.SpanStart);
-                if (csNode is InvocationExpressionSyntax invoke &&
-                    invoke.Expression is MemberAccessExpressionSyntax expr &&
-                    expr.Expression is IdentifierNameSyntax name && name.Identifier.ValueText == "Conversions" &&
-                    expr.Name.Identifier.ValueText == $"To{displayType}") {
-                    return csNode;
-                }
-
-                // Need to use Conversions rather than Convert to match what VB does, eg. True -> -1
-                _extraUsingDirectives.Add("Microsoft.VisualBasic.CompilerServices");
-                var memberAccess = SyntaxFactory.MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
-                    SyntaxFactory.IdentifierName("Conversions"), SyntaxFactory.IdentifierName($"To{displayType}"));
-                var arguments = SyntaxFactory.ArgumentList(SyntaxFactory.SingletonSeparatedList(SyntaxFactory.Argument(csNode)));
-                return SyntaxFactory.InvocationExpression(memberAccess, arguments);
             }
 
             public override CSharpSyntaxNode VisitInvocationExpression(VBSyntax.InvocationExpressionSyntax node)
